lecture-demos/week09/docker-demo/5-environment/requirements.txt
gradio
scikit-learn
joblib
pillow
